feat(web): extract agent registry into dedicated tab
Move EVM and Solana onchain registration out of the Wallet tab
into a new Registry tab, giving each concern its own space.

- Create AgentRegistry.tsx with extracted registration state, API calls, and UI
- Add 6th tab (Registry) to AgentDetail between Logs and Settings
- Clean AgentWallet.tsx to retain only balance, transfers, and history
